change passwordEncoder bean moved to main class, UserDetailsService interface moved to the UserService class, removed roles form User constructor delete UserRoleEnum.java
