+++ src/main/java/com/project/travelAgency/TravelAgencyApplication.java
@@ -2,6 +2,9 @@ package com.project.travelAgency;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 @SpringBootApplication
 public class TravelAgencyApplication {
@@ -9,5 +12,11 @@ public class TravelAgencyApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(TravelAgencyApplication.class, args);
 	}
+
+	@Bean
+	public PasswordEncoder passwordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+
 
 }
