--- src/main/java/com/project/travelAgency/entities/UserRoleEnum.java
@@ -1,22 +1,0 @@
-package com.project.travelAgency.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-
-@AllArgsConstructor
-public enum UserRoleEnum {
-
-    CLIENT("CLIENT"),
-    EMPLOYEE("EMPLOYEE"),
-    ADMIN("ADMIN");
-
-    public String alias;
-
-    @Override
-    public String toString() {
-        return "UserRoleEnum{" +
-                "alias='" + alias + '\'' +
-                '}';
-    }
-}
