+++ src/main/java/com/project/travelAgency/service/UserService.java
@@ -5,6 +5,9 @@ import com.project.travelAgency.entities.UserRole;
 import com.project.travelAgency.exception.NoIdException;
 import com.project.travelAgency.repository.UserRepo;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -12,17 +15,22 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class UserService  {
+public class UserService implements UserDetailsService {
 
     private final PasswordEncoder passwordEncoder;
     private final UserRepo userRepo;
-    //private static final List<UserRole> USER= List.of(new UserRole(1L, "USER"));
+    private static final List<UserRole> USER= List.of(new UserRole(1L, "USER"));
 
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepo.findByUserName(username)
+                .orElseThrow(() -> new NoIdException("User not found"));
+    }
 
     public User save(User user) {
         String encodedPassword = passwordEncoder.encode(user.getPassword());
         user.setPassword(encodedPassword);
-       // user.setUserRoles(USER);
+        user.setUserRoles(USER);
         return userRepo.save(user);
     }
